refactor(2nd-gen): fold version resolution into stamp-since
Delete scripts/resolve-gen2-version.js and inline its version resolution into
scripts/stamp-since.js. Drop the redundant GEN2_VERSION env override; the version
now comes from the 2nd-gen package.json by default, with --version to override and
--from-npm as the transition-only npm `beta` dist-tag bridge. Update the research
and proposal docs to match the single-script shape.
